requirements.txt
streamlit
manim
openai
python-dotenv